requirements.txt
base58==2.0.1
Pillow==7.0.0
git+https://github.com/pklaus/brother_ql.git#egg=brother_ql
qrcode==7.2
Flask==1.1.2
lxml==4.5.0
requests==2.22.0
